requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
PyPDF2==3.0.1
python-docx==0.8.11
flask-vercel==0.4.4